Bump `codeclimate-credo`
The `codeclimate-credo` plugin is currently using an older version of
`credo` and we'd like to update it to the most recent version.

This change pulls a forked version of `credo`. The current
version of `codeclimate-credo` in production is using a fork of `credo`
from the maintainer with tweaks that enable `credo` to work with Code
Climate's spec. Ideally, we'd like to get those changes merged in
upstream, but just preparing this in case we aren't able to.