Refactor
Borrowing some code from: https://github.com/rrrene/credo/pull/592

The maintainer of `credo` brought up a good point on the PR referenced
above, which is that this plugin can work with the output from credo's
JSON output instead of trying to merge a formatter upstream. This is how
most of our plugins work, actually.

This refactor allows us to pull down the latest version of `credo`
instead of maintaining our own version (the previous version was using a
fork of `credo`).